chore: rename Env to Scope and error kinds to align with provider
