+++ packages/adapter-libsql/src/libsql.ts
@@ -83,7 +83,7 @@ class LibSqlQueryable<ClientT extends StdClient | TransactionClient> implements 
       const rawCode = error['rawCode'] ?? e.cause?.['rawCode']
       if (typeof rawCode === 'number') {
         return err({
-          kind: 'Sqlite',
+          kind: 'sqlite',
           extendedCode: rawCode,
           message: error.message,
         })
